chore: organize ui models and components
