+++ app/src/main/java/com/skgtecnologia/sisem/domain/model/bricks/ReportsDetailModel.kt
@@ -3,7 +3,7 @@ package com.skgtecnologia.sisem.domain.model.bricks
 import com.skgtecnologia.sisem.commons.extensions.biLet
 import com.skgtecnologia.sisem.domain.model.header.HeaderModel
 import com.skgtecnologia.sisem.domain.model.header.TextModel
-import com.valkiria.uicomponents.bricks.ReportsDetailUiModel
+import com.valkiria.uicomponents.model.ui.report.ReportsDetailUiModel
 
 data class ReportsDetailModel(
     val header: HeaderModel,
