+++ uicomponents/src/test/java/com/valkiria/uicomponents/bricks/textfield/OutlinedTextFieldViewPaparazziTest.kt
@@ -1,25 +1,25 @@
-package com.valkiria.uicomponents.components.passwordtextfield
+package com.valkiria.uicomponents.bricks.textfield
 
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
-import com.android.ide.common.rendering.api.SessionParams.RenderingMode.V_SCROLL
-import com.valkiria.uicomponents.components.textfield.PasswordTextFieldComponentPreview
+import com.android.ide.common.rendering.api.SessionParams
+import com.valkiria.uicomponents.bricks.textfield.OutlinedTextFieldViewPreview
 import org.junit.Rule
 import org.junit.Test
 
-class PasswordTextFieldComponentPaparazziTest {
+class OutlinedTextFieldViewPaparazziTest {
 
     @get:Rule
     val paparazziRule: Paparazzi = Paparazzi(
         theme = "android:Theme.MaterialComponents.Light.NoActionBar",
         deviceConfig = DeviceConfig.PIXEL_6_PRO.copy(softButtons = false, screenHeight = 1),
-        renderingMode = V_SCROLL
+        renderingMode = SessionParams.RenderingMode.V_SCROLL
     )
 
     @Test
-    fun snapLoginPasswordTextFieldComponent() {
+    fun snapFilledTextFieldView() {
         paparazziRule.snapshot {
-            PasswordTextFieldComponentPreview()
+            OutlinedTextFieldViewPreview()
         }
     }
 }
