+++ uicomponents/src/test/java/com/valkiria/uicomponents/components/chip/FiltersComponentPaparazziTest.kt
@@ -1,4 +1,4 @@
-package com.valkiria.uicomponents.components.filters
+package com.valkiria.uicomponents.components.chip
 
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
