+++ uicomponents/src/test/java/com/valkiria/uicomponents/components/textfield/PasswordTextFieldComponentPaparazziTest.kt
@@ -1,25 +1,25 @@
-package com.valkiria.uicomponents.bricks
+package com.valkiria.uicomponents.components.textfield
 
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
-import com.android.ide.common.rendering.api.SessionParams
-import com.valkiria.uicomponents.bricks.textfield.OutlinedTextFieldViewPreview
+import com.android.ide.common.rendering.api.SessionParams.RenderingMode.V_SCROLL
+import com.valkiria.uicomponents.components.textfield.PasswordTextFieldComponentPreview
 import org.junit.Rule
 import org.junit.Test
 
-class OutlinedTextFieldViewPaparazziTest {
+class PasswordTextFieldComponentPaparazziTest {
 
     @get:Rule
     val paparazziRule: Paparazzi = Paparazzi(
         theme = "android:Theme.MaterialComponents.Light.NoActionBar",
         deviceConfig = DeviceConfig.PIXEL_6_PRO.copy(softButtons = false, screenHeight = 1),
-        renderingMode = SessionParams.RenderingMode.V_SCROLL
+        renderingMode = V_SCROLL
     )
 
     @Test
-    fun snapFilledTextFieldView() {
+    fun snapLoginPasswordTextFieldComponent() {
         paparazziRule.snapshot {
-            OutlinedTextFieldViewPreview()
+            PasswordTextFieldComponentPreview()
         }
     }
 }
